Replaced dangerous heap-based .build() with .sbuild() (renamed to .build()) and minor fixes
Eliminated .build() methods that created Buildable objects on an external (DLL) heap. Renamed .sbuild() methods to .build() (behavior remains unchanged: stack-based object creation).

Applied minor fixes: corrected view-build for objects with std::span handles; fixed the constructor for objects with a specified view (previously, the view was lost); resolved implicit size_t to uint32_t cast warnings using static_cast.

+++ snvoxeng/.src/vk/CommandBuffer.cpp
@@ -110,7 +110,7 @@ void CommandBuffer::cmdCopyImage(
         vkHandle(),
         srcImage, srcImageLayout,
         dstImage, dstImageLayout,
-        regions.size(), regions.data()
+        static_cast<uint32_t>(regions.size()), regions.data()
     );
 }
 void CommandBuffer::cmdCopyImageToBuffer(
